requirements.txt
functions-framework==3.*
polars
google-cloud-bigquery
googlemaps
google-cloud-storage
